Fix: category page 적용 (#47)
* fix: Button props type 수정

* feat: 페이지 개념으로 구조 수정

* feat: 카테고리 페이지 쿼리 적용

---------
