+++ src/api/queries/queryKeys.ts
@@ -1,3 +1,7 @@
 export const userKeys = {
   user: ["user"] as const,
 };
+
+export const categoryKeys = {
+  categories: ["categories"] as const,
+};
