+++ src/api/user/index.ts
@@ -1,5 +1,5 @@
 import { fetcher, fetcherProtected } from "../fetcher";
-import { User } from "./types";
+import { CategoryType, User } from "./types";
 
 // TODO: remove this
 export const getUser = async () => {
@@ -9,4 +9,9 @@ export const getUser = async () => {
 
 export const postSignUp = async (body: FormData) => {
   return await fetcher.post<null>("/users/signup", body);
+};
+
+export const getCategories = async () => {
+  const res = await fetcher.get<CategoryType[]>("/categories");
+  return res.data;
 };
