+++ src/api/user/types.ts
@@ -2,3 +2,9 @@ export type User = {
   nickname: string;
   profileUrl: string;
 };
+
+export type CategoryType = {
+  id: number;
+  title: string;
+  imageUrl: string;
+}
