+++ src/mocks/handlers.ts
@@ -1,5 +1,6 @@
 import { rest } from "msw";
 import { User } from "api/user/types";
+import { categoriesData } from "./data";
 
 export default [
   // TODO: remove this
@@ -13,5 +14,9 @@ export default [
 
   rest.post("/api/users/signin", async (_, res, ctx) => {
     return res(ctx.status(201), ctx.json(null));
+  }),
+
+  rest.get("/api/categories", async (_, res, ctx) => {
+    return res(ctx.status(200), ctx.json(categoriesData));
   }),
 ];
